feat(auth): :sparkles: refactor le service d'authentification en déplaçant la logique dans un nouveau fichier authService et en améliorant la gestion du rafraîchissement des tokens
