+++ src/lib/api.ts
@@ -1,4 +1,3 @@
-import { AuthResponse, LoginCredentials, RefreshTokenResponse, RegisterCredentials } from '@/types/auth';
 import axios from 'axios';
 
 // Créer une instance axios
@@ -35,30 +34,18 @@ api.interceptors.response.use(
       originalRequest._retry = true;
 
       try {
+        // Importation dynamique du service d'authentification pour éviter les dépendances circulaires
+        const authServiceModule = await import('@/services/authService');
+        const authService = authServiceModule.default;
+
         // Tentative de rafraîchissement du token
-        const refreshToken = localStorage.getItem('refreshToken');
-
-        if (!refreshToken) {
-          throw new Error('No refresh token available');
-        }
-
-        const response = await axios.post<RefreshTokenResponse>(
-          `${API_URL}/auth/refresh-token`,
-          { refreshToken }
-        );
-
-        // Mettre à jour le token d'accès
-        const { token: newAccessToken } = response.data.access;
-        localStorage.setItem('accessToken', newAccessToken);
+        const newAccessToken = await authService.refreshToken();
 
         // Mettre à jour le token dans la requête originale et la relancer
         originalRequest.headers.Authorization = `Bearer ${newAccessToken}`;
         return api(originalRequest);
       } catch (refreshError) {
-        // En cas d'échec du rafraîchissement, déconnecter l'utilisateur
-        localStorage.removeItem('accessToken');
-        localStorage.removeItem('refreshToken');
-        window.location.href = '/connexion';
+        // L'erreur sera gérée dans le service d'authentification
         return Promise.reject(refreshError);
       }
     }
@@ -67,74 +54,4 @@ api.interceptors.response.use(
     // par notre utilitaire getErrorMessage
     return Promise.reject(error);
   }
-);
-
-// Service d'authentification
-export const authService = {
-  // Connexion
-  login: async (credentials: LoginCredentials): Promise<AuthResponse> => {
-    try {
-      const response = await api.post<AuthResponse>('/auth/login', credentials);
-
-      // Sauvegarder les tokens
-      localStorage.setItem('accessToken', response.data.accessToken);
-      localStorage.setItem('refreshToken', response.data.refreshToken || response.data.accessToken);
-
-      return response.data;
-    } catch (error) {
-      // Laisser l'erreur se propager pour être traitée par notre gestionnaire d'erreurs
-      throw error;
-    }
-  },
-
-  // Inscription
-  register: async (credentials: RegisterCredentials): Promise<AuthResponse> => {
-    try {
-      const response = await api.post<AuthResponse>('/users', credentials);
-
-      // Sauvegarder les tokens si fournis
-      if (response.data.accessToken) {
-        localStorage.setItem('accessToken', response.data.accessToken);
-        localStorage.setItem('refreshToken', response.data.refreshToken || response.data.accessToken);
-      }
-
-      return response.data;
-    } catch (error) {
-      // Laisser l'erreur se propager pour être traitée par notre gestionnaire d'erreurs
-      throw error;
-    }
-  },
-
-  // Déconnexion
-  logout: async (): Promise<void> => {
-    const refreshToken = localStorage.getItem('refreshToken');
-
-    if (refreshToken) {
-      try {
-        await api.post('/auth/logout', { refreshToken });
-      } catch (error) {
-        console.error('Logout error:', error);
-      }
-    }
-
-    // Nettoyer le stockage local
-    localStorage.removeItem('accessToken');
-    localStorage.removeItem('refreshToken');
-  },
-
-  // Récupérer l'utilisateur courant
-  getCurrentUser: async () => {
-    try {
-      const response = await api.get('/auth/profile');
-      return response.data;
-    } catch (error) {
-      console.error('Get current user error:', error);
-      return null;
-    }
-  },
-
-  // Vérifier si l'utilisateur est connecté
-  isAuthenticated: (): boolean => {
-    return !!localStorage.getItem('accessToken');
-  },
-}; 
+);
